Match output locations from autotools build
This should avoid the need to make any changes to the Principia Makefile
and Azure pipeline config.

The important artifacts are the generated headers and libglog.a. The
autotools build generates the former in src/glog/ and the latter at
.libs/. Since the CMake build uses

    configure_file(<header.in> glog/<header> @ONLY)

to generate the header files, and the output location is relative to
CMAKE_CURRENT_BINARY_DIR, an in-tree build rooted at src/ must be used
to match the header output location from the autotools build. Luckily,
the static library output location is controlled by a separate variable,
so ensuring libglog.a is in the right place is quite straightforward.